@@ -26,7 +26,10 @@
     "@apollo/react-components": "3.1.5",
     "@apollo/react-hooks": "3.1.5",
     "@react-native-community/async-storage": "1.9.0",
+    "@react-native-community/masked-view": "^0.1.10",
     "@react-native-community/netinfo": "5.8.0",
+    "@react-navigation/native": "^5.1.6",
+    "@react-navigation/stack": "^5.2.13",
     "apollo-cache-inmemory": "1.6.5",
     "apollo-cache-persist": "^0.1.1",
     "apollo-client": "2.6.8",
@@ -40,58 +43,60 @@
     "i18n-js": "3.5.1",
     "jsc-android": "241213.1.0",
     "lottie-react-native": "3.3.2",
-    "prop-types": "15.7.2",
+    "prop-types": "^15.7.2",
     "react": "16.13.1",
     "react-native": "0.62.2",
     "react-native-animatable": "1.3.3",
     "react-native-device-info": "5.5.6",
     "react-native-fs": "2.16.6",
-    "react-native-gesture-handler": "1.6.1",
+    "react-native-gesture-handler": "^1.6.1",
     "react-native-google-cast": "^3.3.0",
     "react-native-languages": "3.0.2",
     "react-native-linear-gradient": "2.5.6",
     "react-native-markdown-renderer": "^3.2.8",
     "react-native-orientation": "^3.1.3",
-    "react-native-screens": "2.7.0",
+    "react-native-reanimated": "^1.8.0",
+    "react-native-safe-area-context": "^0.7.3",
+    "react-native-screens": "^2.7.0",
     "react-native-splash-screen": "3.2.0",
     "react-native-static-server": "0.4.2",
     "react-native-vector-icons": "6.6.0",
-    "react-navigation": "4.3.8",
-    "react-navigation-stack": "2.3.13",
     "redux-persist-fs-storage": "^1.3.0",
     "subscriptions-transport-ws": "^0.9.16",
     "update-react-native-app": "1.1.2"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "7.5.5",
-    "@babel/plugin-proposal-decorators": "7.6.0",
-    "@babel/plugin-proposal-export-namespace-from": "7.5.2",
-    "@babel/plugin-proposal-object-rest-spread": "7.6.2",
-    "@babel/plugin-transform-flow-strip-types": "7.6.3",
-    "@babel/plugin-transform-runtime": "7.6.2",
-    "@babel/preset-react": "7.6.3",
-    "@commitlint/cli": "8.2.0",
-    "@commitlint/config-conventional": "8.2.0",
-    "babel-eslint": "10.0.3",
-    "babel-plugin-module-resolver": "3.2.0",
-    "eslint": "6.5.1",
-    "eslint-config-airbnb": "18.0.1",
-    "eslint-config-standard": "14.1.0",
+    "@babel/core": "^7.9.0",
+    "@babel/plugin-proposal-class-properties": "7.8.3",
+    "@babel/plugin-proposal-decorators": "7.8.3",
+    "@babel/plugin-proposal-export-namespace-from": "7.8.3",
+    "@babel/plugin-proposal-object-rest-spread": "7.9.5",
+    "@babel/plugin-transform-flow-strip-types": "7.9.0",
+    "@babel/plugin-transform-runtime": "7.9.0",
+    "@babel/preset-react": "7.9.4",
+    "@babel/runtime": "^7.9.2",
+    "@commitlint/cli": "8.3.5",
+    "@commitlint/config-conventional": "8.3.4",
+    "babel-eslint": "10.1.0",
+    "babel-plugin-module-resolver": "4.0.0",
+    "eslint": "6.8.0",
+    "eslint-config-airbnb": "18.1.0",
+    "eslint-config-standard": "14.1.1",
     "eslint-config-standard-react": "9.2.0",
     "eslint-plugin-babel": "^5.2.1",
-    "eslint-plugin-import": "2.18.2",
-    "eslint-plugin-jest": "22.17.0",
+    "eslint-plugin-import": "2.20.2",
+    "eslint-plugin-jest": "23.8.2",
     "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-node": "10.0.0",
+    "eslint-plugin-node": "11.1.0",
     "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-react": "7.16.0",
+    "eslint-plugin-react": "7.19.0",
     "eslint-plugin-standard": "4.0.1",
     "get-dev-paths": "^0.1.1",
-    "husky": "3.0.9",
-    "jest": "24.9.0",
+    "husky": "4.2.5",
+    "jest": "25.4.0",
     "jest-react-native": "^18.0.0",
-    "metro-react-native-babel-preset": "0.57.0",
-    "react-test-renderer": "16.10.2"
+    "metro-react-native-babel-preset": "0.59.0",
+    "react-test-renderer": "16.13.1"
   },
   "jest": {
     "preset": "react-native"
