@@ -1,6 +1,6 @@
+import 'react-native-gesture-handler'
 import { AppRegistry, YellowBox } from 'react-native'
 import { enableScreens } from 'react-native-screens'
-import 'react-native-gesture-handler'
 
 enableScreens()
 
